B.NowPlaying: Refactoring
There were a number of problems with the NowPlaying extension making its
behaviour, particularly with respect to GUI components, a bit confusing.
In addition there were a number fo hack comments indicating quality
issues. It appears to have suffered some bit-rot as functionality was
added and the extension was expanded.

We address a number of things:

- remove the unused substitute audio display;
- make NowPlayingContents default home in NowPlayingInterface;
- remove all unnecessary graphical programming;
- remove the OnShown timeout hack;
- re-work going fullscreen and back.

Finally we prevent background pattern clearing in the Gdk.Window we ask
GStreamer to draw on, this causes an XLib error on reparenting when the
HeaderBar is in use in the lite interface and the only reason I can find
for clearing the property is the original GStreamer C example code for
the video overlay.