B.WebBrowser: Handle zoom with smooth scrolling
WebView was missing handling for the smooth scrolling direction,
current gdk-sharp doesn't provide the necessary enum values so we add
our own P/Invoked function with the assumption that if we see a 4,
i.e. GDK_SCROLL_SMOOTH [1] then the user has the required version of GDK
for the P/Invoke function we need to calculate the zoom.

[1]: https://developer.gnome.org/gdk3/3.4/gdk3-Event-Structures.html#GDK-SCROLL-SMOOTH:CAPS