B.SoundMenu: Correct Notifications project path
Noticing that the B.SoundMenu project should have referenced the
Notifications library I added its project reference but did it manually,
and more importantly, with the incorrect path in 5784487f8.

Luckily makefile integration was on so users of IDEs that support
that were still using auto-tools and able to build.