browser: Check visiblity when we SetSource
This commit corrects a loading bug for the browser where items were not
displayed on first run. This appears to be linked to the calling the
ClearFilterSelections function so we call that during source setup in
the composite track source contents.

We make one extra change - stop accessing the static ServiceManager in
the check for whether the active source can have a browser or not. This
check doesn't make sense since we hold a reference to the source we're
displaying at the time.